+++ packages/dialtone-mcp-server/src/data.ts
@@ -1,0 +1,13 @@
+// ============================================================================
+// DATA IMPORTS
+// ============================================================================
+
+import utilityClassesData from '@dialpad/dialtone-css/lib/dist/dialtone-docs.json';
+import tokensData from '@dialpad/dialtone-css/lib/dist/tokens-docs.json';
+import componentsData from '@dialpad/dialtone-vue/component-documentation.json';
+import clientRulesData from '../client-rules.json';
+
+export const utilityClasses = utilityClassesData;
+export const tokens = tokensData;
+export const components = componentsData;
+export const clientRules = clientRulesData;
