+++ packages/dialtone-mcp-server/src/index.ts
@@ -2,831 +2,26 @@ import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
 import { CallToolRequestSchema, ListToolsRequestSchema } from "@modelcontextprotocol/sdk/types.js";
 
-import utilityClasses from '@dialpad/dialtone-css/lib/dist/dialtone-docs.json';
-import tokens from '@dialpad/dialtone-css/lib/dist/tokens-docs.json';
-import components from '@dialpad/dialtone-vue/component-documentation.json';
-import clientRules from '../client-rules.json';
+import { utilityClasses, tokens, components, clientRules } from './data.js';
+import { searchUtilityClasses, formatResults, buildCompoundPropertiesSet } from './tools/utility-classes.js';
+import { searchTokens, formatTokenResults } from './tools/tokens.js';
+import { searchComponents, formatComponentResults } from './tools/components.js';
+
+import type {
+  ValueObject,
+  Metadata,
+  ClassData,
+  UtilityClassesData,
+  ThemeData,
+  TokenData,
+  TokensData,
+  ComponentProp,
+  ComponentEvent,
+  ComponentSlot,
+  Component,
+  SearchResult
+} from './types.js';
 
-// ============================================================================
-// TYPE DEFINITIONS
-// ============================================================================
-
-interface ValueObject {
-  prop?: string;
-  value?: string;
-  description?: string;
-}
-
-interface Metadata {
-  deprecated?: boolean;
-  discouraged?: boolean;
-  category?: string;
-  reason?: string;
-  alternatives?: string[];
-  docs?: string;
-  replacement?: string;
-}
-
-interface ClassData {
-  values: ValueObject[];
-  metadata?: Metadata;
-}
-
-interface UtilityClassesData {
-  [className: string]: ClassData;
-}
-
-interface ThemeData {
-  value?: string | number;
-  description?: string;
-}
-
-interface TokenData {
-  [themeName: string]: ThemeData | Metadata;
-}
-
-interface TokensData {
-  [tokenName: string]: TokenData;
-}
-
-interface ComponentProp {
-  name: string;
-  type?: { name: string };
-  values?: string[];
-  description?: string;
-}
-
-interface ComponentEvent {
-  name: string;
-}
-
-interface ComponentSlot {
-  name: string;
-}
-
-interface Component {
-  displayName: string;
-  description?: string;
-  props?: ComponentProp[];
-  events?: ComponentEvent[];
-  slots?: ComponentSlot[];
-  metadata?: Metadata;
-}
-
-interface SearchResult {
-  type: string;
-  name: string;
-  details: any;
-  metadata: Metadata | null;
-  tier?: number;
-}
-
-// ============================================================================
-// SEARCH FUNCTIONS
-// ============================================================================
-
-// Maximum number of results to return (prevents overwhelming responses)
-const MAX_RESULTS = 15;
-
-/**
- * Build a set of all compound properties that exist in the data
- * This is called once when the server starts
- */
-function buildCompoundPropertiesSet(data: UtilityClassesData): Set<string> {
-  const compoundProps = new Set<string>();
-
-  for (const classData of Object.values(data)) {
-    for (const valueObj of classData.values) {
-      const prop = valueObj.prop?.toLowerCase();
-      if (prop && prop.includes('-')) {
-        // Store compound properties like "padding-right", "text-align", etc.
-        compoundProps.add(prop);
-      }
-    }
-  }
-
-  return compoundProps;
-}
-
-/**
- * Extract keywords from a search query
- * Detects compound properties (padding-right) from adjacent words (padding right)
- */
-function extractKeywords(query: string, compoundProperties: Set<string>): { words: string[]; compoundProperties: string[] } {
-  const normalized = query.toLowerCase();
-  const words = normalized.split(/\s+/).filter(w => w.length > 0);
-
-  // Convert px values to rem for Dialtone's 10-based scale
-  const convertedWords = words.flatMap((word: string) => {
-    if (word.endsWith('px')) {
-      const px = parseFloat(word);
-      const rem = `${px / 10}rem`;
-      return [word, rem];
-    }
-    return [word];
-  });
-
-  // Detect compound properties from adjacent words
-  const compoundProps = [];
-  for (let i = 0; i < convertedWords.length - 1; i++) {
-    const word1 = convertedWords[i];
-    const word2 = convertedWords[i + 1];
-
-    // Skip if either word is a number/value
-    if (word1.match(/^\d/) || word2.match(/^\d/)) continue;
-    if (word1.endsWith('px') || word1.endsWith('rem') || word1.endsWith('%')) continue;
-    if (word2.endsWith('px') || word2.endsWith('rem') || word2.endsWith('%')) continue;
-
-    // Check if these two words form a compound property in EITHER order
-    // e.g., "padding right" could be "padding-right" OR "right-padding"
-    const potential1 = `${word1}-${word2}`;
-    const potential2 = `${word2}-${word1}`;
-
-    if (compoundProperties.has(potential1)) {
-      compoundProps.push(potential1);
-    } else if (compoundProperties.has(potential2)) {
-      compoundProps.push(potential2);
-    }
-  }
-
-  return {
-    words: convertedWords,
-    compoundProperties: compoundProps
-  };
-}
-
-/**
- * Check if a keyword is a pure value (not a property name)
- */
-function isValueKeyword(word: string): boolean {
-  // Numeric values with units: 8px, 0.8rem, 50%, 1em
-  if (word.match(/^\d+(\.\d+)?(px|rem|%|em)$/)) return true;
-  // Hex colors: #1C1C1C, #fff
-  if (word.match(/^#[0-9a-f]+$/i)) return true;
-  // Numbers without units: 8, 16, 100
-  if (word.match(/^\d+(\.\d+)?$/)) return true;
-  return false;
-}
-
-/**
- * Helper: Check if a value matches a keyword
- */
-function valueMatchesKeyword(value: string, description: string | undefined, word: string): boolean {
-  // For numeric values with units (8px, 0.8rem, 50%), use token-based matching
-  if (word.match(/^\d+(\.\d+)?(px|rem|%|em)$/)) {
-    const valueTokens = value.split(/\s+/);
-    const descTokens = (description || '').split(/\s+/);
-    return valueTokens.includes(word) || descTokens.includes(word);
-  }
-  // For other keywords, use word boundary matching
-  const escapedWord = word.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-  const wordBoundaryRegex = new RegExp(`\\b${escapedWord}\\b`, 'i');
-  return wordBoundaryRegex.test(value) || wordBoundaryRegex.test(description || '');
-}
-
-/**
- * Smart filter to handle deprecated and discouraged items
- * - Deprecated: Remove completely
- * - Discouraged: Show alternatives from metadata with note
- * - Clean: Keep as-is
- */
-function applySmartFilter(results: SearchResult[], data: UtilityClassesData | TokensData | { [key: string]: Component }): { results: SearchResult[]; notes: string[] } {
-  const filtered: SearchResult[] = [];
-  const notes: string[] = [];
-  const swapped: string[] = [];
-  let deprecatedCount = 0;
-
-  for (const result of results) {
-    // Remove deprecated items completely
-    if (result.metadata?.deprecated) {
-      console.error(`[FILTER] Removing deprecated: ${result.name}`);
-      deprecatedCount++;
-      continue;
-    }
-
-    // Swap discouraged items with alternatives
-    if (result.metadata?.discouraged && result.metadata?.alternatives && result.metadata.alternatives.length > 0) {
-      console.error(`[FILTER] Swapping discouraged: ${result.name} with alternatives: ${result.metadata.alternatives.join(', ')}`);
-
-      // For each alternative pattern
-      for (const altPattern of result.metadata.alternatives) {
-        // Handle wildcard patterns like "d-headline-*"
-        const isWildcard = altPattern.includes('*');
-
-        if (isWildcard) {
-          // Remove wildcard and find matching items
-          const prefix = altPattern.replace(/\*/g, '');
-
-          for (const [name, itemData] of Object.entries(data) as [string, any][]) {
-            if (name.startsWith(prefix)) {
-              // Don't add if already in results or if it's also discouraged/deprecated
-              const alreadyAdded = filtered.some(f => f.name === name) || swapped.includes(name);
-              const isProblematic = itemData.metadata?.deprecated || itemData.metadata?.discouraged;
-
-              if (!alreadyAdded && !isProblematic) {
-                const details = 'values' in itemData
-                  ? { properties: itemData.values }
-                  : result.details;
-                filtered.push({
-                  ...result,
-                  name: name,
-                  details: details,
-                  metadata: itemData.metadata || null
-                });
-                swapped.push(name);
-
-                // Limit alternatives per discouraged item
-                if (swapped.length >= 5) break;
-              }
-            }
-          }
-        } else {
-          // Exact alternative name
-          if (data[altPattern]) {
-            const altData = data[altPattern] as any;
-            const alreadyAdded = filtered.some(f => f.name === altPattern) || swapped.includes(altPattern);
-            const isProblematic = altData.metadata?.deprecated || altData.metadata?.discouraged;
-
-            if (!alreadyAdded && !isProblematic) {
-              const details = 'values' in altData
-                ? { properties: altData.values }
-                : result.details;
-              filtered.push({
-                ...result,
-                name: altPattern,
-                details: details,
-                metadata: altData.metadata || null
-              });
-              swapped.push(altPattern);
-            }
-          }
-        }
-      }
-
-      if (swapped.length > 0) {
-        notes.push(`Replaced discouraged "${result.name}" with recommended alternatives`);
-      }
-      continue;
-    }
-
-    // Keep clean items and discouraged without alternatives
-    filtered.push(result);
-  }
-
-  if (deprecatedCount > 0) {
-    notes.push(`Filtered out ${deprecatedCount} deprecated item${deprecatedCount > 1 ? 's' : ''}`);
-  }
-
-  return { results: filtered, notes };
-}
-
-/**
- * Search utility classes using simple AND-logic (like Dialtone docs site)
- */
-function searchUtilityClasses(query: string, data: UtilityClassesData): { results: SearchResult[]; notes: string[] } {
-  console.error(`\n[CLASS SEARCH DEBUG] Query: "${query}"`);
-
-  // Normalize query: lowercase, replace hyphens/slashes with spaces
-  const normalized = query.toLowerCase().replace(/[/-]/g, ' ');
-  const words = normalized.split(/\s+/).filter(w => w.length > 0);
-
-  // Create regex for each word (handle px/rem conversion)
-  const regexArray = words.map((word: string) => {
-    // For px values, create regex that matches EITHER px OR rem
-    if (word.endsWith('px')) {
-      const px = parseFloat(word);
-      const rem = `${px / 10}rem`;
-      const escapedPx = word.replace(/\./g, '\\.');
-      const escapedRem = rem.replace(/\./g, '\\.');
-      return new RegExp(`(${escapedPx}|${escapedRem})`, 'i');
-    }
-
-    // For rem values, create regex that matches EITHER rem OR px
-    if (word.endsWith('rem')) {
-      const remValue = parseFloat(word);
-      const px = `${remValue * 10}px`;
-      const escapedRem = word.replace(/\./g, '\\.');
-      const escapedPx = px.replace(/\./g, '\\.');
-      return new RegExp(`(${escapedRem}|${escapedPx})`, 'i');
-    }
-
-    // For other words, simple regex
-    const escaped = word.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-    return new RegExp(escaped, 'i');
-  });
-
-  console.error(`[CLASS SEARCH DEBUG] Words:`, words);
-
-  const results: SearchResult[] = [];
-
-  // Check each class: ALL words must match somewhere
-  for (const [className, classData] of Object.entries(data) as [string, ClassData][]) {
-    // Gather all searchable text for this class
-    const searchableTexts = [className.toLowerCase()];
-
-    for (const valueObj of classData.values) {
-      searchableTexts.push(valueObj.prop?.toLowerCase() || '');
-      searchableTexts.push(valueObj.value?.toLowerCase() || '');
-      searchableTexts.push(valueObj.description?.toLowerCase() || '');
-    }
-
-    const combinedText = searchableTexts.join(' ');
-
-    // Check if ALL regexes match somewhere in the combined text
-    const allWordsMatch = regexArray.every((regex: RegExp) => regex.test(combinedText));
-
-    if (allWordsMatch) {
-      results.push({
-        type: 'utility-class',
-        name: className,
-        details: {
-          properties: classData.values
-        },
-        metadata: classData.metadata || null
-      });
-    }
-  }
-
-  console.error(`[CLASS SEARCH DEBUG] Found ${results.length} raw matches`);
-
-  // Apply smart filter (remove deprecated, swap discouraged with alternatives)
-  const { results: filtered, notes } = applySmartFilter(results, data);
-
-  console.error(`[CLASS SEARCH DEBUG] After filter: ${filtered.length} results\n`);
-
-  return { results: filtered, notes };
-}
-
-/**
- * Format search results as readable text
- */
-function formatResults(results: SearchResult[], query: string): string {
-  if (results.length === 0) {
-    return `No results found for "${query}".\n\nTry searching with:\n- Property + value (e.g., "padding right 8px", "display flex")\n- Full property name (e.g., "overflow hidden", "position relative")`;
-  }
-
-  let output = `Found ${results.length} result${results.length > 1 ? 's' : ''} for "${query}":\n\n`;
-
-  results.forEach((result: SearchResult, index: number) => {
-    output += `${index + 1}. **${result.name}**\n`;
-
-    // Show metadata warnings if present
-    if (result.metadata) {
-      if (result.metadata.deprecated) {
-        output += `   ⚠️  **DEPRECATED:** ${result.metadata.reason}\n`;
-      } else if (result.metadata.discouraged) {
-        output += `   ⚠️  **DISCOURAGED:** ${result.metadata.reason}\n`;
-      }
-
-      if (result.metadata.alternatives && result.metadata.alternatives.length > 0) {
-        output += `   📝 **Use instead:** ${result.metadata.alternatives.join(', ')}\n`;
-      }
-
-      if (result.metadata.docs) {
-        output += `   📖 **Docs:** ${result.metadata.docs}\n`;
-      }
-
-      output += `\n`;
-    }
-
-    // Show properties
-    result.details.properties.forEach((prop: ValueObject) => {
-      const desc = prop.description ? ` (${prop.description})` : '';
-      output += `   - ${prop.prop}: ${prop.value}${desc}\n`;
-    });
-
-    // Show usage example
-    output += `   Usage: <div class="${result.name}">...</div>\n\n`;
-  });
-
-  return output;
-}
-
-// ============================================================================
-// TOKEN SEARCH FUNCTIONS
-// ============================================================================
-
-/**
- * Search design tokens using simple AND-logic (like Dialtone docs site)
- */
-function searchTokens(query: string, data: TokensData): { results: SearchResult[]; notes: string[] } {
-  console.error(`\n[TOKEN SEARCH DEBUG] Query: "${query}"`);
-
-  // Normalize query: lowercase, replace hyphens/slashes with spaces
-  const normalized = query.toLowerCase().replace(/[/-]/g, ' ');
-  const words = normalized.split(/\s+/).filter(w => w.length > 0);
-
-  // Create regex for each word (handle px/rem conversion)
-  const regexArray = words.map((word: string) => {
-    // For px values, match EITHER px OR rem
-    if (word.endsWith('px')) {
-      const px = parseFloat(word);
-      const rem = `${px / 10}rem`;
-      const escapedPx = word.replace(/\./g, '\\.');
-      const escapedRem = rem.replace(/\./g, '\\.');
-      return new RegExp(`(${escapedPx}|${escapedRem})`, 'i');
-    }
-
-    // For rem values, match EITHER rem OR px
-    if (word.endsWith('rem')) {
-      const remValue = parseFloat(word);
-      const px = `${remValue * 10}px`;
-      const escapedRem = word.replace(/\./g, '\\.');
-      const escapedPx = px.replace(/\./g, '\\.');
-      return new RegExp(`(${escapedRem}|${escapedPx})`, 'i');
-    }
-
-    // For other words, simple regex
-    const escaped = word.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-    return new RegExp(escaped, 'i');
-  });
-
-  console.error(`[TOKEN SEARCH DEBUG] Words:`, words);
-
-  const results: SearchResult[] = [];
-
-  // Check each token: ALL words must match somewhere
-  for (const [tokenName, themeVariants] of Object.entries(data) as [string, TokenData][]) {
-    // Skip component-specific tokens (not general design tokens)
-    // Only include tokens that start with --dt- or --base-- (like the docs site does)
-    if (!tokenName.startsWith('--dt-') && !tokenName.startsWith('--base--')) {
-      continue;
-    }
-
-    // Gather searchable text: token name + all theme values + descriptions
-    const searchableTexts = [tokenName.toLowerCase()];
-
-    for (const [themeName, themeData] of Object.entries(themeVariants) as [string, ThemeData][]) {
-      if (themeName === 'metadata') continue;
-      if (themeData && themeData.value) {
-        searchableTexts.push(String(themeData.value).toLowerCase());
-      }
-      if (themeData && themeData.description) {
-        searchableTexts.push(String(themeData.description).toLowerCase());
-      }
-    }
-
-    const combinedText = searchableTexts.join(' ');
-
-    // Check if ALL regexes match somewhere
-    const allWordsMatch = regexArray.every((regex: RegExp) => regex.test(combinedText));
-
-    if (allWordsMatch) {
-      const metadata = themeVariants['metadata'] as Metadata | undefined;
-      results.push({
-        type: 'design-token',
-        name: tokenName,
-        details: {
-          allThemes: themeVariants
-        },
-        metadata: metadata || null
-      });
-    }
-  }
-
-  console.error(`[TOKEN SEARCH DEBUG] Found ${results.length} raw matches`);
-
-  // Apply smart filter (remove deprecated, swap discouraged with alternatives)
-  const { results: filtered, notes } = applySmartFilter(results, data);
-
-  console.error(`[TOKEN SEARCH DEBUG] After filter: ${filtered.length} results\n`);
-
-  return { results: filtered, notes };
-}
-
-/**
- * Format token search results with theme variant information
- */
-function formatTokenResults(results: SearchResult[], query: string): string {
-  if (results.length === 0) {
-    return `No token results found for "${query}".\n\nTry searching with:\n- Token category (e.g., "color", "spacing", "border")\n- Specific value (e.g., "#1C1C1C", "0.8rem")\n- Semantic name (e.g., "primary", "success", "danger")`;
-  }
-
-  let output = `Found ${results.length} design token${results.length > 1 ? 's' : ''} for "${query}":\n\n`;
-  output += `ℹ️  **Note:** Design token values change based on the active theme (light/dark mode, brand variant).\n\n`;
-
-  results.forEach((result: SearchResult, index: number) => {
-    output += `${index + 1}. **${result.name}**\n`;
-
-    // Show metadata warnings if present
-    if (result.metadata) {
-      if (result.metadata.deprecated) {
-        output += `   ⚠️  **DEPRECATED:** ${result.metadata.reason}\n`;
-      } else if (result.metadata.discouraged) {
-        output += `   ⚠️  **DISCOURAGED:** ${result.metadata.reason}\n`;
-      }
-
-      if (result.metadata.alternatives && result.metadata.alternatives.length > 0) {
-        output += `   📝 **Use instead:** ${result.metadata.alternatives.join(', ')}\n`;
-      }
-
-      if (result.metadata.docs) {
-        output += `   📖 **Docs:** ${result.metadata.docs}\n`;
-      }
-
-      output += `\n`;
-    }
-
-    // Show theme variants
-    output += `   Theme Variants:\n`;
-    const themes = Object.entries(result.details.allThemes) as [string, ThemeData][];
-
-    // Show first few themes as examples
-    const themesToShow = themes.slice(0, 3);
-    themesToShow.forEach(([themeName, themeData]: [string, ThemeData]) => {
-      const valueStr = themeData && themeData.value ? String(themeData.value) : 'N/A';
-      const descStr = themeData && themeData.description ? String(themeData.description) : '';
-      const desc = descStr ? ` - ${descStr}` : '';
-      output += `   - ${themeName}: ${valueStr}${desc}\n`;
-    });
-
-    if (themes.length > 3) {
-      output += `   - ... and ${themes.length - 3} more theme variants\n`;
-    }
-
-    // Show usage example
-    output += `   Usage: style="color: var(${result.name})"\n`;
-    output += `   Note: This will automatically use the correct value for the active theme.\n\n`;
-  });
-
-  return output;
-}
-
-// ============================================================================
-// COMPONENT SEARCH FUNCTIONS
-// ============================================================================
-
-/**
- * Search Vue components by name, description, or props
- */
-function searchComponents(query: string, components: Component[]): { results: SearchResult[]; notes: string[] } {
-  console.error(`\n[COMPONENT SEARCH DEBUG] Query: "${query}"`);
-
-  // Normalize query: lowercase, replace hyphens/slashes with spaces
-  const normalized = query.toLowerCase().replace(/[/-]/g, ' ');
-  const words = normalized.split(/\s+/).filter(w => w.length > 0);
-
-  // Create regex for each word
-  const regexArray = words.map((word: string) => {
-    const escaped = word.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-    return new RegExp(escaped, 'i');
-  });
-
-  console.error(`[COMPONENT SEARCH DEBUG] Words:`, words);
-
-  const results: SearchResult[] = [];
-
-  // Check each component: ALL words must match somewhere
-  for (const component of components) {
-    // Gather searchable text: name + description + prop names + prop descriptions
-    const searchableTexts = [
-      (component.displayName || '').toLowerCase(),
-      (component.description || '').toLowerCase()
-    ];
-
-    for (const prop of component.props || []) {
-      searchableTexts.push((prop.name || '').toLowerCase());
-      searchableTexts.push((prop.description || '').toLowerCase());
-    }
-
-    const combinedText = searchableTexts.join(' ');
-
-    // Check if ALL regexes match somewhere
-    const allWordsMatch = regexArray.every((regex: RegExp) => regex.test(combinedText));
-
-    if (allWordsMatch) {
-      results.push({
-        type: 'component',
-        name: component.displayName,
-        details: {
-          description: component.description,
-          props: component.props || [],
-          events: component.events || [],
-          slots: component.slots || []
-        },
-        metadata: component.metadata || null
-      });
-    }
-  }
-
-  console.error(`[COMPONENT SEARCH DEBUG] Found ${results.length} raw matches`);
-
-  // Apply smart filter (remove deprecated, swap discouraged with alternatives)
-  // Note: Pass components array as data source for potential alternatives
-  const componentsData: { [key: string]: Component } = {};
-  components.forEach((c: Component) => { componentsData[c.displayName] = c; });
-  const { results: filtered, notes } = applySmartFilter(results, componentsData);
-
-  console.error(`[COMPONENT SEARCH DEBUG] After filter: ${filtered.length} results\n`);
-
-  return { results: filtered, notes };
-}
-
-/**
- * Format component search results
- */
-function formatComponentResults(results: SearchResult[], query: string): string {
-  if (results.length === 0) {
-    return `No components found for "${query}".\n\nTry searching with:\n- Component name (e.g., "button", "modal", "avatar")\n- Partial name (e.g., "badge", "card")\n- Component feature (e.g., "dropdown", "tooltip")`;
-  }
-
-  let output = `Found ${results.length} component${results.length > 1 ? 's' : ''} for "${query}":\n\n`;
-
-  results.forEach((result: SearchResult, index: number) => {
-    output += `${index + 1}. **${result.name}**\n`;
-
-    // Show metadata warnings if present
-    if (result.metadata) {
-      if (result.metadata.deprecated) {
-        output += `   ⚠️  **DEPRECATED:** ${result.metadata.reason}\n`;
-      }
-
-      if (result.metadata.replacement) {
-        output += `   📝 **Use instead:** ${result.metadata.replacement}\n`;
-      }
-
-      if (result.metadata.docs) {
-        output += `   📖 **Docs:** ${result.metadata.docs}\n`;
-      }
-
-      output += `\n`;
-    }
-
-    // Show description
-    if (result.details.description) {
-      output += `   ${result.details.description}\n\n`;
-    }
-
-    // Show key props (up to 5)
-    if (result.details.props && result.details.props.length > 0) {
-      output += `   **Key Props:**\n`;
-      const propsToShow = result.details.props.slice(0, 5);
-      propsToShow.forEach((prop: ComponentProp) => {
-        const typeName = prop.type?.name || 'unknown';
-        const values = prop.values ? ` Options: ${prop.values.slice(0, 3).join(', ')}${prop.values.length > 3 ? '...' : ''}` : '';
-        output += `   • \`${prop.name}\` (${typeName})${values}\n`;
-        if (prop.description) {
-          output += `     ${prop.description}\n`;
-        }
-      });
-
-      if (result.details.props.length > 5) {
-        output += `   • ... and ${result.details.props.length - 5} more props\n`;
-      }
-      output += `\n`;
-    }
-
-    // Show events if any
-    if (result.details.events && result.details.events.length > 0) {
-      output += `   **Events:** ${result.details.events.map((e: ComponentEvent) => e.name).slice(0, 3).join(', ')}`;
-      if (result.details.events.length > 3) {
-        output += ` (${result.details.events.length - 3} more)`;
-      }
-      output += `\n\n`;
-    }
-
-    // Show slots if any
-    if (result.details.slots && result.details.slots.length > 0) {
-      output += `   **Slots:** ${result.details.slots.map((s: ComponentSlot) => s.name).slice(0, 3).join(', ')}`;
-      if (result.details.slots.length > 3) {
-        output += ` (${result.details.slots.length - 3} more)`;
-      }
-      output += `\n\n`;
-    }
-
-    // Usage example
-    output += `   **Usage:**\n`;
-    output += `   \`\`\`vue\n`;
-    output += `   import { ${result.name} } from '@dialpad/dialtone-vue'\n`;
-    output += `   \`\`\`\n\n`;
-  });
-
-  return output;
-}
-
-/**
- * Sort unified results by tier, then metadata status, then name
- * Priority: Tier 1 > Tier 2, Clean > Discouraged > Deprecated
- */
-function sortUnifiedResults(results: SearchResult[]): SearchResult[] {
-  return results.sort((a: SearchResult, b: SearchResult) => {
-    // First sort by tier (1 comes before 2)
-    if ((a.tier || 0) !== (b.tier || 0)) {
-      return (a.tier || 0) - (b.tier || 0);
-    }
-
-    // Within same tier, sort by metadata status
-    const getMetadataScore = (result: SearchResult): number => {
-      if (!result.metadata) return 0; // Clean (no metadata) = highest priority
-      if (result.metadata.deprecated) return 2; // Deprecated = lowest priority
-      if (result.metadata.discouraged) return 1; // Discouraged = middle priority
-      return 0; // Clean
-    };
-
-    const aScore = getMetadataScore(a);
-    const bScore = getMetadataScore(b);
-
-    if (aScore !== bScore) {
-      return aScore - bScore;
-    }
-
-    // Finally sort alphabetically by name
-    return a.name.localeCompare(b.name);
-  });
-}
-
-/**
- * Format unified results (all types in one list, no tiers)
- */
-function formatUnifiedResults(results: SearchResult[], query: string): string {
-  if (results.length === 0) {
-    return `No results found for "${query}".\n\nTry searching with:\n- CSS properties (e.g., "padding", "display", "flex")\n- Values (e.g., "8px", "0.8rem", "100%")\n- Token names (e.g., "color", "spacing", "border")\n- Component names (e.g., "button", "modal", "avatar")\n- Hex colors (e.g., "#1C1C1C")`;
-  }
-
-  let output = `Found ${results.length} result${results.length > 1 ? 's' : ''} for "${query}":\n\n`;
-
-  // Format all results in order
-  results.forEach((result: SearchResult, index: number) => {
-    output += formatSingleResult(result, index + 1);
-  });
-
-  return output;
-}
-
-/**
- * Format a single result (utility class, token, or component)
- */
-function formatSingleResult(result: SearchResult, index: number): string {
-  let output = `${index}. **${result.name}** _(${result.type})_\n`;
-
-  // Show metadata warnings if present
-  if (result.metadata) {
-    if (result.metadata.deprecated) {
-      output += `   ⚠️  **DEPRECATED:** ${result.metadata.reason}\n`;
-    }
-    if (result.metadata.discouraged) {
-      output += `   ⚠️  **DISCOURAGED:** ${result.metadata.reason || 'Consider alternatives'}\n`;
-    }
-    if (result.metadata.replacement) {
-      output += `   📝 **Use instead:** ${result.metadata.replacement}\n`;
-    }
-    if (result.metadata.docs) {
-      output += `   📖 **Docs:** ${result.metadata.docs}\n`;
-    }
-    output += `\n`;
-  }
-
-  // Format based on type
-  if (result.type === 'utility-class') {
-    // Show properties
-    if (result.details.properties && result.details.properties.length > 0) {
-      result.details.properties.forEach((prop: ValueObject) => {
-        output += `   • \`${prop.prop}\`: \`${prop.value}\``;
-        if (prop.description) {
-          output += ` (${prop.description})`;
-        }
-        output += `\n`;
-      });
-    }
-    output += `\n`;
-  } else if (result.type === 'design-token') {
-    // Show token values (first theme only for brevity)
-    if (result.details.allThemes) {
-      const themes = Object.keys(result.details.allThemes);
-      if (themes.length > 0) {
-        const firstTheme = themes[0];
-        const themeData = result.details.allThemes[firstTheme];
-        output += `   • Value: \`${themeData.value}\``;
-        if (themeData.description) {
-          output += ` (${themeData.description})`;
-        }
-        output += `\n`;
-        if (themes.length > 1) {
-          output += `   • Available in ${themes.length} theme${themes.length > 1 ? 's' : ''}\n`;
-        }
-      }
-    }
-    output += `\n`;
-  } else if (result.type === 'component') {
-    // Show description
-    if (result.details.description) {
-      output += `   ${result.details.description}\n\n`;
-    }
-
-    // Show key props (up to 3 for brevity in unified view)
-    if (result.details.props && result.details.props.length > 0) {
-      output += `   **Key Props:** `;
-      const propsToShow = result.details.props.slice(0, 3);
-      output += propsToShow.map((p: ComponentProp) => `\`${p.name}\``).join(', ');
-      if (result.details.props.length > 3) {
-        output += ` (+${result.details.props.length - 3} more)`;
-      }
-      output += `\n\n`;
-    }
-  }
-
-  return output;
-}
 
 async function main() {
   // Build compound properties set from utility classes data (done once at startup)
